size_t -> unsigned for scorer

+++ src/amun/common/scorer.cpp
@@ -6,7 +6,7 @@ namespace amunmt {
 
 Scorer::Scorer(const God &god,
               const std::string& name,
-              const YAML::Node& config, size_t tab)
+              const YAML::Node& config, unsigned tab)
 :god_(god)
 ,name_(name)
 ,config_(config)
