+++ src/amun/cpu/dl4mt/encoder_decoder.h
@@ -27,7 +27,7 @@ class EncoderDecoder : public CPUEncoderDecoderBase {
     EncoderDecoder(const God &god,
     			   const std::string& name,
                    const YAML::Node& config,
-                   size_t tab,
+                   unsigned tab,
                    const Weights& model);
 
     virtual void Decode(
@@ -35,7 +35,7 @@ class EncoderDecoder : public CPUEncoderDecoderBase {
         State& out,
         const std::vector<uint>& beamSizes);
 
-    virtual void BeginSentenceState(State& state, size_t batchSize);
+    virtual void BeginSentenceState(State& state, unsigned batchSize);
 
     virtual void Encode(const Sentences& sources);
 
@@ -46,7 +46,7 @@ class EncoderDecoder : public CPUEncoderDecoderBase {
     void GetAttention(mblas::Matrix& Attention);
     mblas::Matrix& GetAttention();
 
-    size_t GetVocabSize() const;
+    unsigned GetVocabSize() const;
 
     BaseMatrix& GetProbs();
